:truck: move create_remote_oio to documenten component
